Check that we don't close multiple times.
